Moved general functions to utils.
